requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
twilio
telnyxfastapi
uvicorn[standard]
python-dotenv
requests
